+++ src/data/en/index.ts
@@ -52,7 +52,8 @@ export const instanceData = {
       threads: 'Threads',
       topic: 'Topic',
       subject: 'Subject',
-      userProfile: 'User Profile'
+      userProfile: 'User Profile',
+      privacyPolicy: 'Privacy Policy',
     },
     roles: {
       donor: "Donor",
@@ -90,9 +91,16 @@ export const instanceData = {
       chooseOption: 'Click on one of the options.',
       trashedNotice: 'This content is marked for deletion.'
     },
+    consent: {
+      title: 'Consent for external Content',
+      intro: 'While using this site you may allowed us to load content from external providers. You can read about the details in the %privacypolicy%.',
+      revokeTitle: 'Revoke',
+      revokeText: 'Here you can revoke your consent. In this case we ask again, before we load content from those providers',
+      noConsent: 'No content saved.',
+      revokeConsent: 'Revoke consent',
+    },
     embed: {
       text: 'By clicking on image or button above you agree that external content from %provider% will be loaded. Also personal data may be transferred to this service in accordance with our %privacypolicy%.',
-      link: 'Privacy Policy',
       video: 'Play Video',
       applet: 'Load Applet',
       twingle: 'Load Donation Form'
