+++ src/data/en/menu-data.ts
@@ -69,6 +69,10 @@ export const footerData: InstanceData['footerData'] = {
       children: [
         { title: 'Privacy Policy', url: `https://de.${serloDomain}/privacy` },
         {
+          title: 'Revoke consent',
+          url: `/consent`,
+        },
+        {
           title: 'Terms of use',
           url: `https://de.${serloDomain}/terms`,
         },
